Split flow classes into their own files
