+++ src/flow/conditional.js
@@ -1,0 +1,18 @@
+class Conditional {
+  constructor(redirector, condition) {
+    this.redirector = redirector;
+    this.condition = condition;
+  }
+
+  redirect(req, res) {
+    if (this.condition()) {
+      this.redirector.redirect(req, res);
+    }
+  }
+
+  check() {
+    return this.condition();
+  }
+}
+
+module.exports = Conditional;
